feat: 경기 시작하는 부분 분리
